refactor(web): P3 — carve wave 2 (knowledge_mem, skills_api, schedules_api)
RFC-003 P3. Five sections move verbatim into three domain modules as
per-section registrars at their original composition positions:
knowledge (incl. the versioning sub-block), memory notes, learned
context -> knowledge_mem.py; skills -> skills_api.py; schedules ->
schedules_api.py. __init__ is down to 1,937 lines.

The carve script now carries the wave-1 lesson as executable
assertions: section boundaries stop at bare registrar CALL lines (not
just the next marker), every register_* def must be called exactly once
from the composition root, and no composition may hide inside a domain
module. The boundary assertion fired on the first pass of this very
wave — the schedules section had swallowed fifteen wave-1 call lines —
and the re-carve landed clean.

One import restored verbatim (croniter, unconditional like the
original); datetime/croniter dropped from __init__ where they became
unused. Gate: new=0. Suite 6,079 green; parity 183 exact-order.
